+++ pkg/providers/gtm/testdata/TestDataGtmResources/missing_domain.tf
@@ -1,0 +1,6 @@
+provider "akamai" {
+  edgerc = "../../test/edgerc"
+}
+
+data "akamai_gtm_resources" "my_gtm_resources" {
+}